requirements.txt
absl-py==0.11.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
fsspec==0.8.4
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
importlib-metadata==2.0.0
Markdown==3.3.3
numpy~=1.19.2
oauthlib==3.1.0
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytorch-lightning==1.0.5
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
torch~=1.6.0
torchvision~=0.7.0
tqdm==4.51.0
urllib3==1.25.11
Werkzeug==1.0.1
zipp==3.4.0

opencv-python~=4.4.0.46
Flask~=1.1.2
matplotlib~=3.3.3
PyJWT~=2.0.0